ref(issueDetails): move DetectorDetails type to a leaf module
views/issueDetails/context imported the DetectorDetails type from the
detectorSection component, while detectorSection imports the issue
details context back — a 2-cycle. DetectorDetails is a pure interface,
so moving it to a sibling leaf module (and importing it from there)
removes the context from the type-import cycle.

Effect: largest type-import SCC 2193 -> 2106.
