+++ static/app/views/issueDetails/sidebar/detectorDetails.tsx
@@ -1,0 +1,11 @@
+export interface DetectorDetails {
+  description?: string;
+  detectorId?: string;
+  detectorPath?: string;
+  detectorSlug?: string;
+  detectorType?:
+    | 'metric_alert'
+    | 'cron_monitor'
+    | 'uptime_monitor'
+    | 'mobile_build_monitor';
+}
