+++ static/app/views/issueDetails/sidebar/detectorSection.tsx
@@ -16,19 +16,8 @@ import {makeAlertsPathname} from 'sentry/views/alerts/pathnames';
 import {useDetectorQuery} from 'sentry/views/detectors/hooks';
 import {makeMonitorDetailsPathname} from 'sentry/views/detectors/pathnames';
 import {useIssueDetails} from 'sentry/views/issueDetails/context';
+import type {DetectorDetails} from 'sentry/views/issueDetails/sidebar/detectorDetails';
 import {SidebarSectionTitle} from 'sentry/views/issueDetails/sidebar/sidebar';
-
-export interface DetectorDetails {
-  description?: string;
-  detectorId?: string;
-  detectorPath?: string;
-  detectorSlug?: string;
-  detectorType?:
-    | 'metric_alert'
-    | 'cron_monitor'
-    | 'uptime_monitor'
-    | 'mobile_build_monitor';
-}
 
 export function getDetectorDetails({
   event,
